pyproject.toml
[project]
name = "kembang-ai"
version = "0.1.0"
description = "Agency-first AI chatbot SaaS for UMKM customer service automation"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    # Cache
    "redis>=5.0.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Settings
    "pydantic-settings>=2.2.0",
    "email-validator>=2.1.0",
    # LLM / AI
    "httpx>=0.27.0",
    "litellm>=1.52.0",
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    # RAG
    "rank-bm25>=0.2.2",
    # File handling
    "python-multipart>=0.0.9",
    "pymupdf>=1.27.1",
    "reportlab>=4.4.10",
    "fpdf2>=2.8.7",
    "sentence-transformers>=5.2.3",
    "cachetools>=7.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
